+++ test/helpers/expectBNIsWithinDelta.js
@@ -1,0 +1,8 @@
+const expectBNIsWithinDelta = (actualBN, expected, allowableDelta) => {
+  const delta = actualBN.minus(expected).abs().toNumber();
+  expect(delta).to.be.lt(allowableDelta);
+};
+
+module.exports = {
+  expectBNIsWithinDelta,
+};